Migrate all of Model to BigDecimal

+++ model/src/main/java/ar/edu/itba/paw/model/StockMarketEntry.java
@@ -3,6 +3,7 @@ package ar.edu.itba.paw.model;
 import org.jetbrains.annotations.NotNull;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Table(name = "stockmarket")
 @Entity
@@ -14,7 +15,7 @@ public class StockMarketEntry {
     @Column(name = "resourcetype")
     private int _resourcetype;
 
-    private double amount;
+    private BigDecimal amount;
 
     @PostLoad
     private void postLoad() {
@@ -23,7 +24,7 @@ public class StockMarketEntry {
 
     public StockMarketEntry(){}
 
-    public StockMarketEntry(@NotNull ResourceType resourceType, double amount) {
+    public StockMarketEntry(@NotNull ResourceType resourceType, BigDecimal amount) {
         this.resourceType = resourceType;
         this.amount = amount;
         _resourcetype = resourceType.getId();
@@ -33,11 +34,11 @@ public class StockMarketEntry {
         return resourceType;
     }
 
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
 }
