+++ model/src/main/java/ar/edu/itba/paw/model/packages/Formatter.java
@@ -1,8 +1,10 @@
 package ar.edu.itba.paw.model.packages;
 
+import java.math.BigDecimal;
+
 /**
  * Created by juanfra on 08/04/17.
  */
 public interface Formatter {
-    String format(Double d);
+    String format(BigDecimal d);
 }
